+++ src/EventSubscriber/GetAvailableResourceClassActionsEventSubscriber.php
@@ -10,18 +10,10 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class GetAvailableResourceClassActionsEventSubscriber implements EventSubscriberInterface
 {
-    private const AVAILABLE_FORM_ITEM_ACTIONS = [
-        AuthorizationService::READ_FORM_ACTION,
-        AuthorizationService::UPDATE_FORM_ACTION,
-        AuthorizationService::DELETE_FORM_ACTION,
-        AuthorizationService::CREATE_SUBMISSIONS_FORM_ACTION,
-        AuthorizationService::READ_SUBMISSIONS_FORM_ACTION,
-        AuthorizationService::UPDATE_SUBMISSIONS_FORM_ACTION,
-        AuthorizationService::DELETE_SUBMISSIONS_FORM_ACTION,
-    ];
+    private const AVAILABLE_FORM_ITEM_ACTIONS = AuthorizationService::FORM_ITEM_ACTIONS;
     private const AVAILABLE_FORM_COLLECTION_ACTIONS = [AuthorizationService::CREATE_FORMS_ACTION];
 
-    private const AVAILABLE_SUBMISSION_ITEM_ACTIONS = [];
+    private const AVAILABLE_SUBMISSION_ITEM_ACTIONS = AuthorizationService::SUBMISSION_ITEM_ACTIONS;
     private const AVAILABLE_SUBMISSION_COLLECTION_ACTIONS = [];
 
     public static function getSubscribedEvents(): array